Allow for other statuses (e.g. UNKNOWN, WARNING) to be set
By adding a 3rd parameter, we can set the status directly, instead of it being determined by the healthy flag on the first param.

For backwards compatibility I've left the other params as-is, but it could be reduced to just a status and errorMessage check. Any non-OK status should be considered unhealthy.